// projects/ui/src/lib/types/alignement.ts
export type JustifyContent =
  | 'normal'
  | 'stretch'
  | 'flex-start'
  | 'flex-end'
  | 'center'
  | 'space-between'
  | 'space-around'
  | 'space-evenly'
  | 'start'
  | 'end'
  | 'left'
  | 'right'
  | 'safe center'
  | 'unsafe center'
  | 'inherit'
  | 'initial'
  | 'revert'
  | 'revert-layer'
  | 'unset';

export const ALL_JUSTIFY_CONTENT: JustifyContent[] = [
  'normal',
  'stretch',
  'flex-start',
  'flex-end',
  'center',
  'space-between',
  'space-around',
  'space-evenly',
  'start',
  'end',
  'left',
  'right',
  'safe center',
  'unsafe center',
  'inherit',
  'initial',
  'revert',
  'revert-layer',
  'unset',
];

export type AlignItems =
  | 'normal'
  | 'stretch'
  | 'flex-start'
  | 'flex-end'
  | 'center'
  | 'baseline'
  | 'first baseline'
  | 'last baseline'
  | 'start'
  | 'end'
  | 'self-start'
  | 'self-end'
  | 'anchor-center'
  | 'safe center'
  | 'unsafe center'
  | 'inherit'
  | 'initial'
  | 'revert'
  | 'revert-layer'
  | 'unset';

export const ALL_ALIGN_ITEMS: AlignItems[] = [
  'normal',
  'stretch',
  'flex-start',
  'flex-end',
  'center',
  'baseline',
  'first baseline',
  'last baseline',
  'start',
  'end',
  'self-start',
  'self-end',
  'anchor-center',
  'safe center',
  'unsafe center',
  'inherit',
  'initial',
  'revert',
  'revert-layer',
  'unset',
];